+++ src/editor/view/viewComponent/viewComponent.ts
@@ -1,0 +1,22 @@
+import { IDisposable } from "src/base/common/dispose";
+import { IEditorViewModel } from "src/editor/common/viewModel";
+import { EditorViewEventHandler } from "src/editor/view/editorViewHandler";
+
+
+/**
+ * @class // TODO
+ */
+ export class EditorViewComponent extends EditorViewEventHandler {
+
+    public readonly viewModel: IEditorViewModel;
+
+    constructor(viewModel: IEditorViewModel) {
+        super();
+        this.viewModel = viewModel;
+    }
+
+    public addViewComponent(id: string, component: EditorViewComponent): IDisposable {
+        return this.viewModel.addViewComponent(id, component);
+    }
+
+}
